Separate file for WriteBatch
